Adds Java8 method refactor and RESTController nature,code to interface
