+++ jcart-site/src/main/resources/application-default.properties
@@ -4,7 +4,7 @@ spring.profiles.active=dev
 
 ################### Web Configuration ##########################
 
-#spring.thymeleaf.mode=LEGACYHTML5
+spring.thymeleaf.mode=HTML
 #spring.thymeleaf.cache=false
 
 server.port=8443
