+++ jcart-site/src/main/resources/templates/myAccount.html
@@ -54,7 +54,7 @@
 						</div>
 						<div role="tabpanel" class="tab-pane fade" id="orders">
 							<h2>Orders</h2>
-							<table cellspacing="0" class="shop_table cart">
+							<table class="shop_table cart">
 								<thead>
 									<tr>
 										<th>#</th>
